+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `tintcolor` of group layers is now loaded
 - It is now possible to lookup if a tile was flipped:
 	`GID::flip_horizontal()`, `GID::flip_vertical()`, `GID::flip_diagonal()`
+- Support loading custom properties
 
 ### Changed
 - Most structs with public fields are now marked `#[non_exhaustive]`